linter and outsource code
