+++ src/services/user/index.js
@@ -1,45 +1,18 @@
-const errors = require('@feathersjs/errors');
 const service = require('feathers-mongoose');
-const user = require('./model');
+const { userModel, registrationPinModel } = require('./model');
 const hooks = require('./hooks');
 const registrationPinsHooks = require('./hooks/registrationPins');
 const publicTeachersHooks = require('./hooks/publicTeachers');
 const firstLoginHooks = require('./hooks/firstLogin');
-const { AdminUsers } = require('./services');
+const { AdminUsers, UserLinkImportService } = require('./services');
 const adminHook = require('./hooks/admin');
 
-const userDataFilter = user => ({
-	userId: user._id,
-	email: user.email,
-	firstName: user.firstName,
-	lastName: user.lastName,
-	importHash: user.importHash,
-	schoolId: user.schoolId,
-	birthday: user.birthday,
-});
-
-class UserLinkImportService {
-	constructor(userService) {
-		this.userService = userService;
-		this.docs = {};
-	}
-
-	get(hash, params) { // can not use get becouse the hash can have / that mapped to non existing routes
-		return this.userService.find({ query: { importHash: hash } })
-			.then((users) => {
-				if (users.data.length !== 1) {
-					throw new errors.BadRequest('Can not match the hash.');
-				}
-				return userDataFilter(users.data[0]);
-			}).catch(err => err);
-	}
-}
 
 module.exports = function setup() {
 	const app = this;
 
 	const options = {
-		Model: user.userModel,
+		Model: userModel,
 		paginate: {
 			default: 1000,
 			max: 1000,
@@ -56,7 +29,7 @@ module.exports = function setup() {
 
 	/* publicTeachers Service */
 	app.use('/publicTeachers', service({
-		Model: user.userModel,
+		Model: userModel,
 		paginate: {
 			default: 25,
 			max: 1000,
@@ -70,7 +43,7 @@ module.exports = function setup() {
 
 	/* registrationPin Service */
 	app.use('/registrationPins', service({
-		Model: user.registrationPinModel,
+		Model: registrationPinModel,
 		paginate: {
 			default: 500,
 			max: 5000,
