+++ src/services/user/services/index.js
@@ -1,5 +1,7 @@
 const AdminUsers = require('./AdminUsers');
+const UserLinkImportService = require('./UserLinkImportService');
 
 module.exports = {
 	AdminUsers,
+	UserLinkImportService,
 };
